+++ src/wallet/wallet_tx_builder.cpp
@@ -50,10 +50,7 @@ PrepareTransactionResult WalletTxBuilder::PrepareTransaction(
                 switch (rtype) {
                     case ReceiverType::P2PKH:
                     case ReceiverType::P2SH:
-                        // TODO: This is the correct policy, but it’s a breaking change from
-                        //       previous behavior, so enable it separately. (#6409)
-                        // if (strategy.AllowRevealedRecipients()) {
-                        if (!spendable.utxos.empty() || strategy.AllowRevealedRecipients()) {
+                        if (strategy.AllowRevealedRecipients()) {
                             result.insert(OutputPool::Transparent);
                         }
                         break;
@@ -428,10 +425,7 @@ PrivacyPolicy TransactionEffects::GetRequiredPrivacyPolicy() const
     if (!spendable.utxos.empty()) {
         // TODO: Add a check for whether we need AllowLinkingAccountAddresses here. (#6467)
         if (payments.HasTransparentRecipient()) {
-            // TODO: AllowFullyTransparent is the correct policy, but it’s a breaking change from
-            //       previous behavior, so enable it separately. (#6409)
-            // maxPrivacy = PrivacyPolicy::AllowFullyTransparent;
-            return PrivacyPolicy::AllowRevealedSenders;
+            return PrivacyPolicy::AllowFullyTransparent;
         } else {
             return PrivacyPolicy::AllowRevealedSenders;
         }